Refactoring and lulz etc. - Outputs .phuzz files with results

+++ phuzz/__init__.py
@@ -6,6 +6,8 @@ import re
 import os
 import logging
 import socket
+import pickle
+from hashlib import md5
 from random import randint
 from base64 import b32encode
 from tempfile import mkstemp
@@ -23,20 +25,21 @@ FUNCALL_RE = re.compile(r'^-> ((?P<cls>[^\-]+)->)?(?P<fnc>[^\s\(]+)' +
                         r'\((?P<args>.*?)\)$')
 # C-style quoted and escaped string
 CSTRING_RE = (r'(?P<quo>[\'"])' +
-              r'(?P<val>(\\.|[^(?P=quo)]+)*)' +
+              r'(?P<val>(\\.|[^(?P=quo)])*)' +
               r'(?P=quo)')
 # Separate agruments to functions, from xdebug trace log
-CALLARGS_RE = re.compile(r'(?P<args>(^|\s*,\s*)?(' + r'(?P<str>' + CSTRING_RE +
-                         r'|[^,\)]+' + r')))')
-SYSLOG_RE = re.compile(
-    '^(\[[^\]]*\]\s+)?' + '(?P<fun>[^\(]+)' + '\((?P<args>.*?)\)' +
-    '\s*=\s*(?P<ret>[^\'" ]+)(\s+[^\'"]+)?$')
+CALLARGS_RE = re.compile(r'(?P<args>(^|\s*,\s*)(' + r'(?P<str>'
+                         + CSTRING_RE + r'|[^,\)]+' + r')))', re.MULTILINE)
+STRACE_RE = re.compile(
+    r'^(\[[^\]]*\]\s+)?' + r'(?P<fun>[^\(]+)' + r'\((?P<args>.*?)\)' +
+    r'\s*=\s*(?P<ret>[^\'" ]+)(\s+[^\'"]+)?$')
 
 
 Loc = namedtuple('Loc', ['file', 'line'])
 LogMessage = namedtuple('LogMessage', ['msg', 'loc'])
 Var = namedtuple('Var', ['name', 'key', 'value', 'loc'])
 Func = namedtuple('Func', ['fun', 'args', 'loc'])
+PhuzzCase = namedtuple('PhuzzCase', ['url', 'root', 'state', 'traces'])
 
 
 def unescape(data):
@@ -50,32 +53,33 @@ def unlink(*args):
         if arg and os.path.exists(arg):
             try:
                 os.unlink(arg)
-            except:
+            except Exception:
                 pass
 
 
-def snapshot(*files):
+def snapshot(*files, **kwa):
     ret = []
     for filename in files:
         data = None
         try:
             if os.path.exists(filename):
-                with open(filename, "r+") as fh:
-                    data = fh.read()
-                    fh.truncate(0)
-        except:
+                with open(filename, "r+") as handle:
+                    data = handle.read()
+                    handle.truncate(0)
+        except Exception:
             pass
-        unlink(filename)
+        if kwa.get('remove'):
+            unlink(filename)
         ret.append(data)
     return ret
 
 
-def try_connect(listen):
+def try_connect(addr):
     try:
-        sock = socket.create_connection(listen, 1)
+        sock = socket.create_connection(addr, 1)
         sock.close()
         return True
-    except:
+    except Exception:
         return False
 
 
@@ -86,12 +90,12 @@ def parse_logs(regex, data):
         for match in regex.finditer(data)]
 
 
-def parse_syslog(data, ignore_files=None):
+def parse_strace(data, ignore_files=None):
     ret = []
     if not data:
         return ret
     for entry in data.split("\n"):
-        match = SYSLOG_RE.match(entry)
+        match = STRACE_RE.match(entry)
         if match:
             mdat = match.groupdict()
             ignored = [X in mdat['args'] for X in ignore_files]
@@ -107,43 +111,68 @@ def calls_scan_vars(entries):
         if len(entry.fun) == 2 and entry.fun[0] in PHP_GLOBALS])
 
 
+class DoubleHasher(object):
+    def __init__(self):
+        self.hasher = md5()
+
+    def update(self, val):
+        """Prevents concatenation from resulting in identical hashes"""
+        self.hasher.update(str(val))
+        self.hasher.update(self.hasher.digest())
+
+    def hexdigest(self):
+        return self.hasher.hexdigest()
+
+
 class SyscallTracer(object):
     def __init__(self, target):
         self.target = target
         self.proc = None
         self.logfile = mkstemp('strace')[1]
         self.logfh = None
+        self.cmd = None
         self._setup()
 
     def _setup(self):
+        # TODO: implement dtruss, dtrace, systemtap etc.
+        self.cmd = ' '.join([
+            'exec', 'strace', '-qyfy', '-s', '4096',
+            '-p', str(self.target.pid), '2>', self.logfile
+        ])
         yama_ptrace_scope = '/proc/sys/kernel/yama/ptrace_scope'
         if os.getuid() != 0 and os.path.exists(yama_ptrace_scope):
-            with open(yama_ptrace_scope, 'r') as fh:
-                if fh.read() != "0\n":
+            with open(yama_ptrace_scope, 'r') as handle:
+                if handle.read() != "0\n":
                     print("On Linux, this will only work after you run:")
                     print("  echo 0 | sudo tee " + yama_ptrace_scope)
                     raise RuntimeError(yama_ptrace_scope + " must be 0")
 
-    def begin(self):
-        self.finish()
-        # TODO: implement dtruss (OSX) and dtrace (possibly?)
-        cmd = ['strace', '-qyfy', '-s', '4096', '-p', str(self.target.pid)]
+    def start(self):
+        self.stop()
         self.logfh = open(self.logfile, "w")
-        self.proc = subprocess.Popen(cmd, universal_newlines=True,
-                                     stderr=self.logfh)
+        self.proc = subprocess.Popen(self.cmd, universal_newlines=True, shell=True)
         if self.proc.poll() is not None:
-            raise RuntimeError("Could not strace: " + str(cmd))
+            raise RuntimeError("Could not strace: " + str(self.cmd))
+        LOG.info('SyscallTracer started, pid: %r', self.proc.pid)
 
-    def finish(self):
+    def stop(self):
         if self.proc:
-            subprocess.Popen(['kill', '-9', str(self.proc.pid)])
-            self.proc.wait()
+            LOG.debug('SyscallTracer stopped, pid: %r', self.proc.pid)
+            self.proc.terminate()
             self.proc = None
-            return snapshot(self.logfile)[0]
+        if self.logfh:
+            self.logfh.close()
+            self.logfh = None
+
+    def reset(self):
+        self.logfh.truncate(0)
+
+    def snapshot(self):
+        return snapshot(*(self.logfile,), remove=False)[0]
 
 
 class PHPHarness(object):
-    def __init__(self, listen, root, preload=None, ini={}):
+    def __init__(self, listen, root, preload=None, ini=None):
         self._check_php_modules()
         self.proc = None
         self.listen = listen
@@ -152,6 +181,8 @@ class PHPHarness(object):
         self.strace = None
         self.xdebug_path = mkstemp('.xdebug')[1]
         self.logs = [self.xdebug_path + '.xt']
+        if ini is None:
+            ini = dict()
         if preload:
             ini["auto_prepend_file"] = preload
         self.ini = self._config(ini)
@@ -204,22 +235,24 @@ class PHPHarness(object):
             LOG.debug("Waiting for server...")
             time.sleep(0.5)
         self.strace = SyscallTracer(self.proc)
+        self.strace.start()
 
     def stop(self):
         if self.proc:
             self.proc.terminate()
+            self.proc = None
         if self.strace:
-            self.strace.finish()
+            self.strace.stop()
             self.strace = None
         unlink(*self.logs)
 
     def trace_begin(self):
         snapshot(*self.logs)
-        self.strace.begin()
+        self.strace.reset()
 
     def trace_finish(self):
         logdata = snapshot(*self.logs)
-        strace_out = self.strace.finish()
+        strace_out = self.strace.snapshot()
         logdata.append(strace_out)
         return logdata
 
@@ -250,11 +283,74 @@ class Trace(object):
         return out
 
 
-class Analyzer(object):
-    __slots__ = ('php', 'interwebs')
+class CaseManager(object):
+    def __init__(self, output_dir):
+        self.output = output_dir
 
-    def __init__(self, php):
+    def _is_interesting(self, case):
+        for _, trace in case.traces.items():
+            for entry in trace:
+                if entry[0]:
+                    return True
+
+    def _hash_traces(self, traces):
+        hasher = DoubleHasher()
+        for subsys, trace in traces.items():
+            hasher.update(subsys)
+            for varsused, entry in trace:
+                # Arguments to the function are ignored as they change often
+                hasher.update(entry.fun)
+                if varsused:
+                    for what, name in varsused:
+                        hasher.update(what)
+                        hasher.update(name)
+                loc = entry.loc
+                if loc:
+                    hasher.update(loc.file)
+                    hasher.update(loc.line)
+        return hasher.hexdigest()
+
+    def _display(self, case):
+        overview = {}
+        for subsys, call_list in case.traces.items():
+            highlights = []
+            loc = None
+            for values, call in call_list:
+                if not values:
+                    continue
+                args = call.args
+                if not isinstance(args, list):
+                    args = [args]
+                if loc is None or (call.loc and loc.file != call.loc.file):
+                    loc = call.loc
+                    if loc:
+                        highlights.append(loc.file)
+                    highlights.append(' '.join(["\t", '->'.join(call.fun),
+                                                "(", ', '.join(args), ")"]))
+            if highlights:
+                overview[subsys] = highlights
+        if overview:
+            for subsys, highlights in overview.items():
+                print(subsys)
+                print("\n".join(highlights))
+            print()
+
+    def ingest(self, case):
+        if not self._is_interesting(case):
+            return False
+        digest = self._hash_traces(case.traces)
+        filename = os.path.join(self.output, digest + '.phuzz')
+        if not os.path.exists(filename):
+            with open(filename, 'wb') as handle:
+                pickle.dump(case, handle)
+        self._display(case)
+
+class Phuzzer(object):
+    __slots__ = ('php', 'interwebs', 'manager')
+
+    def __init__(self, php, manager):
         self.php = php
+        self.manager = manager
         interwebs = requests.Session()
         interwebs.max_redirects = 0
         interwebs.max_retries = 0
@@ -266,15 +362,16 @@ class Analyzer(object):
         xdebug = filter(lambda L: L.loc.file != self.php.preload,
                         parse_logs(TRACELOG_RE, xdebug))
         ignore_files = ['2</dev/pts'] + self.php.logs
-        return Trace(resp, xdebug, parse_syslog(strace, ignore_files))
+        return Trace(resp, xdebug, parse_strace(strace, ignore_files))
 
     def _request_for_state(self, url, state):
         params = state['_REQUEST']
         params.update(state['_GET'])
-        method = state.get('METHOD', 'POST' if any(['_POST' in state, '_FILES' in state]) else 'GET')
+        needs_post = any(['_POST' in state, '_FILES' in state])
+        method = state.get('METHOD', 'POST' if needs_post else 'GET')
         return self.interwebs.request(method, url, params=params,
-                                cookies=state['_COOKIE'],
-                                allow_redirects=False)
+                                      cookies=state['_COOKIE'],
+                                      allow_redirects=False)
 
     def trace(self, url, state=None):
         if state is None:
@@ -286,39 +383,30 @@ class Analyzer(object):
             return self._collect(resp)
         self.php.trace_finish()
 
+    def _scan_calls(self, calls, state):
+        return [(self._scan_call(call, state), call)
+                for call in calls]
+
     def _scan_call(self, call, state):
+        """Annotates calls with which values from input state were found"""
         if call.fun[0] in PHP_GLOBALS:
-            return
-        for val in state.values():
-            for subval in val.values():
+            return None  # Exclude these?
+        found = []
+        for topkey, val in state.items():
+            for subkey, subval in val.items():
                 args = call.args
                 if not isinstance(args, list):
                     args = [args]
                 for arg_val in args:
                     if subval in arg_val:
-                        return ' '.join([
-                            "\t", '->'.join(call.fun),
-                            "(", ', '.join(args), ")"])
+                        found.append((topkey, subkey))
+        return found
 
-    def _scan(self, state, phpcalls, syscalls):
-        loc = None
-        php_highlights = []
-        for call in phpcalls:
-            res = self._scan_call(call, state)
-            if res:
-                if loc is None or loc.file != call.loc.file:
-                    loc = call.loc
-                    php_highlights.append(self.php.translate_path(loc.file))
-                php_highlights.append(res)
-        php_highlights = filter(None, php_highlights)
-        sys_calls = filter(None, [self._scan_call(call, state)
-                                  for call in syscalls])
-        if len(php_highlights):
-            print("\n".join(php_highlights))
-        if len(sys_calls):
-            print("\nsyscalls:")
-            print("\n".join(sys_calls))
-        print()
+    def _scan(self, url, state, phpcalls, syscalls):
+        phpcalls = self._scan_calls(phpcalls, state)
+        syscalls = self._scan_calls(syscalls, state)
+        traces = dict(php=phpcalls, sys=syscalls)
+        return PhuzzCase(url, self.php.root, state, traces)
 
     def run_file(self, filepath):
         webpath = filepath[len(self.php.root):]
@@ -350,4 +438,6 @@ class Analyzer(object):
                 for key, val in newvars:
                     state[key][val] = b32encode(os.urandom(randint(1, 4) * 5))
                 new_states = len(newvars) > 0
-            self._scan(state, phpcalls, trace.syslog)
+            result = self._scan(url, state, phpcalls, trace.syslog)
+            if result:
+                self.manager.ingest(result)
